v1.5.3

+++ uploady/profile/actions/delete_me.php
@@ -1,6 +1,9 @@
 <?php
 include_once '../../session.php';
 
-$user->deleteUser($_SESSION['username']);
-
-$utils->redirect($utils->siteUrl('/logout.php'));
+if ($auth->checkToken($_GET['token'], $_SESSION['csrf'])) {
+    $user->deleteUser($_SESSION['username']);
+    $utils->redirect($utils->siteUrl('/logout.php?redirect=delete'));
+} else {
+    $utils->redirect($utils->siteUrl('profile/account.php?msg=csrf'));
+}
